Make a hook page
The anon bump page looks for the user's "own_tag" cookie and, if found,
it simply redirects to the non-anon bump page.

If there is no stored own_tag then it displays its content. This allows
the anonymous user to either create an own_tag and proceed to the
non-anon bump, or just go directly to the shared meme URL without any
further interaction with the DendriteNetwork.

In order to make the process seem a little less clunky I'm going to
create a simpler page that redirects to the real anonbump page if there
isn't a (cookie-)stored own_tag.

The better solution would be to just simplify the anon bump page and have
it do the heavy lifting only after it was determined there was no own_tag
stored on the browser.  Maybe I'll do that later.